+++ apps/explorer/lib/explorer/application.ex
@@ -128,7 +128,8 @@ defmodule Explorer.Application do
         configure(Explorer.Chain.Cache.RootstockLockedBTC),
         configure(Explorer.Migrator.TransactionsDenormalization),
         configure(Explorer.Migrator.AddressCurrentTokenBalanceTokenType),
-        configure(Explorer.Migrator.AddressTokenBalanceTokenType)
+        configure(Explorer.Migrator.AddressTokenBalanceTokenType),
+        configure(Explorer.Migrator.SanitizeMissingBlockRanges)
       ]
       |> List.flatten()