+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- [#9360](https://github.com/blockscout/blockscout/pull/9360) - Move missing ranges sanitize to a separate background migration
+
 ### Fixes
 
 ### Chore